+++ cli/uc-cmd.py
@@ -1640,7 +1640,7 @@ _pd_cols = {
     'list_lpi_mods': ['mod', 'logic', 'description', 'version', 'api'],
     'modhelp_lpi': ['name', 'type', 'required', 'default', 'help'],
     'modhelp_phi': ['name', 'type', 'required', 'default', 'help'],
-    'list_datapullers': ['name', 'active', 'pid', 'cmd'],
+    'list_datapullers': ['name', 'active', 'pid', 'state', 'cmd'],
 }
 
 _fancy_indentsp = {